requirements.txt
gradio
pandas
scikit-learn
PyPDF2
python-dotenv
psycopg2-binary
google-genai
matplotlib